+++ src/routes/tags/+page.server.ts
@@ -1,12 +1,13 @@
-import { getAllPostsMeta } from '$lib/content';
+import { getAllTags } from '$lib/content';
 import { parseLanguage } from '$lib/utils/language';
 import type { PageServerLoad } from './$types';
 
 export const load: PageServerLoad = async ({ url }) => {
 	const lang = parseLanguage(url);
 	return {
-		posts: getAllPostsMeta(lang),
+		tags: getAllTags(lang),
 		lang
 	};
 };
+